+++ src/curios_comic_crawler/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = 'Luraminaki'
 __email__ = ''
-__version__ = '0.3.0'
+__version__ = '0.4.0'